check for null app before trying to kill it

+++ jacs2-services/src/main/java/org/janelia/jacs2/asyncservice/spark/SparkAppProcessor.java
@@ -170,7 +170,7 @@ public class SparkAppProcessor extends AbstractSparkProcessor<Void> {
                                                 runningClusterState.getData().getSparkClusterInfo().getMasterURI(),
                                                 runningClusterState.getData().getSparkClusterInfo().getMasterJobId(),
                                                 runningClusterState.getData().getSparkClusterInfo().getWorkerJobId())));
-                        sparkApp.kill(); // terminate the app just in case it is still running
+                        if (sparkApp != null) sparkApp.kill(); // terminate the app just in case it is still running
                         if (exc != null) {
                             logger.error("Spark processing error encountered", exc);
                             jacsServiceDataPersistence.updateServiceState(
